feat(grouped_by): refactor charge model validation
