+++ app/services/charges/validators/base_service.rb
@@ -14,6 +14,8 @@ module Charges
       def valid?
         # NOTE: override and add validation rules
 
+        validate_grouped_by
+
         if errors?
           result.validation_failure!(errors:)
           return false
@@ -27,6 +29,17 @@ module Charges
       private
 
       attr_reader :charge
+
+      def grouped_by
+        properties["grouped_by"]
+      end
+
+      def validate_grouped_by
+        return if grouped_by.nil? || grouped_by.is_a?(Array) && grouped_by.blank?
+        return if grouped_by.is_a?(Array) && grouped_by.all? { |f| f.is_a?(String) } && grouped_by.all?(&:present?)
+
+        add_error(field: :grouped_by, error_code: "invalid_type")
+      end
     end
   end
 end
